requirements.txt
django-taggit==1.3.0
Django==3.1.7
django-extensions==3.1.1
py==1.10.0
pylint==2.7.2
Pillow==8.1.2
sorl-thumbnail==12.7.0
pymorphy2==0.9.1
